+++ app/Calculation/MemberGroupStatistic.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace App\Calculation;
+
+class MemberGroupStatistic
+{
+    /**
+     * memberIdList
+     * @var int[]
+     */
+    public array $in = [];
+
+    /**
+     * memberIdList
+     * @var int[]
+     */
+    public array $unsure = [];
+
+    /**
+     * memberIdList
+     * @var int[]
+     */
+    public array $out = [];
+
+    /**
+     * memberIdList
+     * @var int[]
+     */
+    public array $attended = [];
+}